+++ inst/doc/Custom_Office_Templates.R
@@ -35,39 +35,39 @@ ex_yaml_read = yaml::read_yaml(ex_yaml_tmpfile)
 # b64_example_layout_docx_png         = knitrdata::data_encode(if_example_layout_docx_png       , encoding="base64")
 
 ## ----eval=FALSE---------------------------------------------------------------
-#  library(onbrand)
-#  file.copy(system.file(package="onbrand","examples","example.pptx"), ".", overwrite = TRUE)
-#  file.copy(system.file(package="onbrand","examples","example.docx"), ".", overwrite = TRUE)
-#  file.copy(system.file(package="onbrand","examples","example.yaml"), ".", overwrite = TRUE)
+# library(onbrand)
+# file.copy(system.file(package="onbrand","examples","example.pptx"), ".", overwrite = TRUE)
+# file.copy(system.file(package="onbrand","examples","example.docx"), ".", overwrite = TRUE)
+# file.copy(system.file(package="onbrand","examples","example.yaml"), ".", overwrite = TRUE)
 
 ## ----echo=FALSE, comment='', message=TRUE, eval=TRUE--------------------------
 cat(readLines(ex_yaml_tmpfile) , sep="\n")
 
 ## ----eval=FALSE---------------------------------------------------------------
-#  library(onbrand)
-#  vlres = view_layout(template    = "example.pptx",
-#                      output_file = "example_layout.pptx")
+# library(onbrand)
+# vlres = view_layout(template    = "example.pptx",
+#                     output_file = "example_layout.pptx")
 
 ## ----eval=FALSE---------------------------------------------------------------
-#   obnd = read_template(template = "example.pptx",
-#                        mapping  = "example.yaml")
-
-## ----eval=FALSE---------------------------------------------------------------
-#  obnd = preview_template(obnd)
-#  save_report(obnd, "example_preview.pptx")
-
-## ----eval=FALSE---------------------------------------------------------------
-#  library(onbrand)
-#  vlres = view_layout(template    = "example.docx",
-#                      output_file = "example_layout.docx")
-
-## ----eval=FALSE---------------------------------------------------------------
-#  obnd = read_template(template = "example.docx",
+#  obnd = read_template(template = "example.pptx",
 #                       mapping  = "example.yaml")
 
 ## ----eval=FALSE---------------------------------------------------------------
-#  obnd = preview_template(obnd)
-#  save_report(obnd, "example_preview.docx")
+# obnd = preview_template(obnd)
+# save_report(obnd, "example_preview.pptx")
+
+## ----eval=FALSE---------------------------------------------------------------
+# library(onbrand)
+# vlres = view_layout(template    = "example.docx",
+#                     output_file = "example_layout.docx")
+
+## ----eval=FALSE---------------------------------------------------------------
+# obnd = read_template(template = "example.docx",
+#                      mapping  = "example.yaml")
+
+## ----eval=FALSE---------------------------------------------------------------
+# obnd = preview_template(obnd)
+# save_report(obnd, "example_preview.docx")
 
 ## ----echo=FALSE, comment='', message=TRUE, eval=TRUE--------------------------
 cat(readLines(file.path(system.file(package="onbrand"), "examples", "example.yaml")) , sep="\n")